pyproject.toml
[project]
name = "ai-quiz-builder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "bilalinit", email = "ahmedbilalyt20@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.6.1",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "utils>=1.0.2",
]

[project.scripts]
ai-quiz-builder = "ai_quiz_builder:main"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"
